chore: seperate nodejs server config
